+++ src/utils.js
@@ -37,18 +37,3 @@ export const getNodeURL = () => {
   cachedNodeURL = getConfig('NODE_URL', getRandomItem(ETH_NODES));
   return cachedNodeURL;
 }
-
-export const TOKEN = {
-  USDT: {
-    addr: '0xdac17f958d2ee523a2206206994597c13d831ec7',
-    decimals: 6,
-  },
-  DAI: {
-    addr: '0x6B175474E89094C44Da98b954EedeAC495271d0F',
-    decimals: 18,
-  },
-  USDC: {
-    addr: '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48',
-    decimals: 6,
-  }
-}